Remove null constraint on Option#result
It's probably going to be easier if we allow ourselves to enter the
"basics" of an encounter first, which would mean relaxing the null
constraints. This will make it easier to rapidly build out the forms.

We can always have messages on an Encounter edit form that indicate what
is missing, and have badges on the index or something.